Split test into set of tests and move code from test_counting_iterator into main function to be closer to initial test state
